Reorganizing code locations and testing autoboxing version of ProtoLongArray
