+++ src/main/proto/packedatlas/ProtoBoxedLongArray.proto
@@ -1,0 +1,10 @@
+syntax = "proto2";
+
+option java_multiple_files = true;
+option java_outer_classname = "ProtoBoxedLongArrayWrapper";
+
+package org.openstreetmap.atlas.proto;
+
+message ProtoBoxedLongArray {
+    repeated int64 value = 1;
+}
