+++ src/main/proto/packedatlas/ProtoPrimitiveLongArray.proto
@@ -1,7 +1,7 @@
 syntax = "proto2";
 
 option java_multiple_files = true;
-option java_outer_classname = "PrimitiveLongArrayWrapper";
+option java_outer_classname = "ProtoPrimitiveLongArrayWrapper";
 
 package org.openstreetmap.atlas.proto;
