+++ src/test/java/org/openstreetmap/atlas/proto/converters/ProtoTagListConverterTest.java
@@ -1,4 +1,4 @@
-package org.openstreetmap.atlas.geography.converters.proto;
+package org.openstreetmap.atlas.proto.converters;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -8,6 +8,7 @@ import java.util.Map;
 import org.junit.Assert;
 import org.junit.Test;
 import org.openstreetmap.atlas.proto.ProtoTag;
+import org.openstreetmap.atlas.proto.converters.ProtoTagListConverter;
 
 /**
  * @author lcram
